# Conflicts:
#	backend/src/app/fill_db/create_tuples/types.json
#	backend/src/app/routers/cityManager/city_checker.py
#	backend/src/app/routers/cityManager/router.py
#	frontend/src/Game/PlanetViewer/CityViewer/CityManager.js
#	frontend/src/Game/PlanetViewer/CityViewer/Grids/CityInfoGrid.js

+++ backend/src/app/routers/generalRouter/schemas.py
@@ -9,4 +9,4 @@ class GeneralModifiersScheme(BaseModel):
     stat: str
     modifier: float
     political_stance: str
-    political_stance_modifier: int
+    political_stance_modifier: float
